Changed and organized project structure, files and folders
